refactor : S3 이미지 관련 로직 리팩토링
